Nice clean up, much easier to work with
